Add execution of tasks according to the requirements of hw-03
The logic for working with the form has been added to the RegistrationScreen and LoginScreen components. When the form is submitted, data is collected from the fields and displayed in the console and Alert. Additionally, the logic for automatically dismissing the keyboard when clicking outside the form using Keyboard.dismiss has been implemented.